Reorder utils code
