+++ backend/apps/accounts/admin.py
@@ -1,5 +1,41 @@
 from django.contrib import admin
+from django.contrib.auth import REDIRECT_FIELD_NAME
+from django.contrib.auth.decorators import user_passes_test
 
 from .modelss import User
 
 admin.site.register(User)
+
+
+def student_required(function=None, redirect_field_name=REDIRECT_FIELD_NAME, login_url="login"):
+    """
+    Redirect unauthenticated students to login page.
+
+    Decorator for views that checks that the logged in user is a student,
+    redirects to the log-in page if necessary.
+    """
+    actual_decorator = user_passes_test(
+        lambda u: u.is_active and u.is_student,
+        login_url=login_url,
+        redirect_field_name=redirect_field_name,
+    )
+    if function:
+        return actual_decorator(function)
+    return actual_decorator
+
+
+def hospital_required(function=None, redirect_field_name=REDIRECT_FIELD_NAME, login_url="login"):
+    """
+    Redirect unauthenticated hospitals to login page.
+
+    Decorator for views that checks that the logged in user is a hospital,
+    redirects to the log-in page if necessary.
+    """
+    actual_decorator = user_passes_test(
+        lambda u: u.is_active and u.is_hospital,
+        login_url=login_url,
+        redirect_field_name=redirect_field_name,
+    )
+    if function:
+        return actual_decorator(function)
+    return actual_decorator
