+++ backend/apps/accounts/src/notifications.py
@@ -1,32 +1,10 @@
 # adapted from: https://gist.github.com/jcinis/2866253
 import logging
-from random import choice
-from string import ascii_lowercase, digits
 
 from django.conf import settings
 from django.contrib.auth.forms import PasswordResetForm
 
-from .modelss import User
-
 logger = logging.getLogger("django")
-
-
-def generate_random_username(length=16, chars=ascii_lowercase + digits, split=4, delimiter="-"):
-
-    username = "".join([choice(chars) for i in range(length)])
-
-    if split:
-        username = delimiter.join(
-            [username[start : start + split] for start in range(0, len(username), split)]
-        )
-
-    try:
-        User.objects.get(username=username)
-        return generate_random_username(
-            length=length, chars=chars, split=split, delimiter=delimiter
-        )
-    except User.DoesNotExist:
-        return username
 
 
 def send_password_set_email(
