+++ backend/apps/ineedstudent/views/hospital_posting_edit.py
@@ -6,7 +6,7 @@ from django.utils.decorators import method_decorator
 from django.utils.translation import gettext as _
 from django.views.generic.edit import FormView
 
-from apps.accounts.decorator import hospital_required
+from apps.accounts.admin import hospital_required
 from apps.ineedstudent.forms import PostingForm
 
 logger = logging.getLogger(__name__)
